Separate code in CardIO handler.
